make navbar optional
